requirements.txt
dash>=2.17.1
dash_core_components>=2.0.0
dash_html_components>=2.0.0
paho_mqtt>=2.1.0
pandas>=2.2.2
plotly>=5.23.0
dash-bootstrap-components>=1.6.0
sqlalchemy>=2.0.32